Dynamic pressure data file
I decided to go ahead and upload this file to simplify the instructions for how to compile the code. The file is 24 MB, which is larger than I would like to include with the source code, but as long as it's only this one file, I think it is okay for now.